Bind BackendController as a global controller
